add setting to force highlight colors to be distinct from each other
i had to loosen the "distinction" criteria for background colors because
otherwise it created a situation where requesting a high CR *and*
distinct hilight colors meant that literally nothing was getting
through. this means that hilight colors can be a bit dim on darker
themes, but ultimately it should work out.